+++ web-wallet/CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Fix UI not scrolling to top after wizard and sub-route navigation [#2997]
 - Fix edge case in Dusk to Lux conversion [#3032]
+- Fix inconsistent terminology usage for "Mnemonic Phrase" [#3035]
 
 ## [0.8.0] - 2024-11-19
 
@@ -367,6 +368,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [#3010]: https://github.com/dusk-network/rusk/issues/3010
 [#3028]: https://github.com/dusk-network/rusk/issues/3028
 [#3032]: https://github.com/dusk-network/rusk/issues/3032
+[#3035]: https://github.com/dusk-network/rusk/issues/3035
 
 <!-- VERSIONS -->